Modularize the parser
